+++ src/main/java/com/quasiris/qsf/util/UrlUtil.java
@@ -1,0 +1,8 @@
+package com.quasiris.qsf.util;
+
+public class UrlUtil {
+
+    public static String removePassword(String url) {
+        return url.replaceAll("(.*)://(.*)@(.*)", "$1://$3");
+    }
+}